+++ src/common/services/UserState/index.ts
@@ -1,0 +1,5 @@
+import { UserState } from "./UserState";
+import { useUserStore } from "./zustand/zustand";
+
+export { useUserStore };
+export type { UserState };
