--- src/common/services/UserStore/index.ts
@@ -1,5 +1,0 @@
-import { UserState } from "./UserState/UserState";
-import { useUserStore } from "./UserStore";
-
-export { useUserStore };
-export type { UserState };
